重构直播模块：迁移 Web API 并修复弹幕连接
- 新增 LiveRoomAPI，集中封装分区、推荐、关注、房间、排行、礼物和勋章接口，统一 Web Cookie、Wbi 与 CSRF 参数
- 直播首页、分区、用户中心和直播间迁移至现行 Web 响应结构，补齐 HTTPS 图片、错误处理及 long 类型 UID 适配
- 播放地址切换至 getRoomPlayInfo，解析多线路、格式、编码和清晰度信息，并同步桌面与移动直播页切流逻辑
- 弹幕连接改用 getDanmuInfo 动态节点与 token，按协议 v2 完成鉴权、心跳、分包和 zlib 解压
- 发送弹幕及礼物改用 Web 登录态和 bili_jct，Cookie 读取切换至 HTTPS 以包含 Secure 凭据
- 清理已失效的小视频、旧榜单和旧签到入口，更新直播导航及 README 完成状态

+++ BiliBili.UWP/Models/LiveModels.cs
@@ -273,7 +273,7 @@ namespace BiliBili.UWP.Models
 
 
         public string type { get; set; }
-        public int uid { get; set; }
+        public long uid { get; set; }
         public int short_id { get; set; }
         public string tags { get; set; }
         public int online { get; set; }
